:truck: api: move get-certification-details usecase from lib to src
